requirements.txt
streamlit>=1.34
pandas>=2.2
openpyxl
lseg.data
datetime
refinitiv.data